Added fan and corrected light
